Add file list as an input to dcp2
This patch adds the support for giving a file list as an input
of dcp2. Dcp2 reads the input file and generates a flist. It
then stats all the files/directories in the flist, and check
whether the files/directories are under any source paths. Only
the files/directories that locate under one of the source paths
will be copied into destination directory. Other files/directories
will be ignored.